+++ ui/src/app/components/runtime-configuration-panel/create-panel.spec.tsx
@@ -8,7 +8,6 @@ import { render, screen, waitFor } from '@testing-library/react';
 import { toAnalysisConfig } from 'app/utils/analysis-config';
 import { ComputeType } from 'app/utils/machines';
 import { runtimePresets } from 'app/utils/runtime-presets';
-import { PanelContent } from 'app/utils/runtime-utils';
 import { serverConfigStore } from 'app/utils/stores';
 
 import defaultServerConfig from 'testing/default-server-config';
@@ -21,6 +20,7 @@ import {
 import { buildWorkspaceStub } from 'testing/stubs/workspaces';
 
 import { CreatePanel, CreatePanelProps } from './create-panel';
+import { PanelContent } from './utils';
 
 const defaultAnalysisConfig = toAnalysisConfig(
   defaultGceRuntimeWithPd(),
