+++ ui/src/app/components/runtime-configuration-panel/utils.tsx
@@ -1,0 +1,11 @@
+export enum PanelContent {
+  Create = 'Create',
+  Customize = 'Customize',
+  DeleteRuntime = 'DeleteRuntime',
+  DeleteUnattachedPd = 'DeleteUnattachedPd',
+  DeleteUnattachedPdAndCreate = 'DeleteUnattachedPdAndCreate',
+  Disabled = 'Disabled',
+  ConfirmUpdate = 'ConfirmUpdate',
+  ConfirmUpdateWithDiskDelete = 'ConfirmUpdateWithDiskDelete',
+  SparkConsole = 'SparkConsole',
+}
